correct sqrt

+++ src/snapchat/Sqrt.java
@@ -12,12 +12,12 @@ public class Sqrt {
             long mid = l + (r - l) / 2;
             long sqrt = mid * mid;
             if (sqrt == x)
-                return (long)mid;
+                return (int)mid;
             if (sqrt < x) 
                 l = mid + 1;
             else
                 r = mid - 1;
         }
-        return (long)r;
+        return (int)r;
     }
 }
